+++ src/components/payment/sellerInfo.module.scss
