+++ src/pages/idea-market/IdeaMarketPayment.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import PaymentTitle from '../../components/payment/PaymentTitle';
-import SellerInfo from '../../components/info/SellerInfo';
+import SellerInfo from '../../components/payment/SellerInfo';
 import PaymentMethods from '../../components/payment/PaymentMethods';
 import PaymentSummary from '../../components/payment/PaymentSummary';
-import PaymentButton from '../../components/common/button/PaymentButton';
+import PaymentButton from '../../components/payment/PaymentButton';
 import styles from './ideaMarketPayment.module.scss';
 
 const IdeaMarketPayment: React.FC = () => {
